Internal API

This page lists unexported functionality from KdotP, that may be of interest to developers.

Unexported, internal functionality

KdotP.minimal_generators_indices — Method

Returns a set of indices into group g which generates the group g, i.e., such that

idxs = minimal_generators_indices(g; cntr)
g′ = generate(g[idxs]; cntr)

where g′ ~ g (but may differ by trivial reciprocal lattice vectors in some operations).

KdotP.poormans_sparsification — Method

Poor man's "matrix sparsification" via the reduced row echelon form.