KdotP.jl

API

KdotP.kdotp — Function

kdotp(lgir::LGIrrep,
      αβγ=nothing;
      timereversal::Bool = true,
      degree::Union{Integer, Nothing} = nothing)  -->  Vector{MonomialHamiltonian}

Return a basis for the allowed k⋅p Hamiltonians of a small irrep lgir up to linear degree in momentum.

Input arguments

lgir: a small irrep of a little group, provided as an instance of the type LGIrrep. Tables of LGIrreps are accessible from the lgirreps function of Crystalline.jl package.

αβγ: if the little group associated with lgir has free parameters, i.e., if the k-point is parametrized by free parameters (α, β, γ), these parameters may be set via αβγ = [α, β, γ]. If αβγ = nothing (default), lgir is implicitly evaluated at αβγ = [0, 0, 0].

Keyword arguments

timereversal (keyword argument): if true (default), time-reversal invariance is imposed on the k⋅p expansion. If lgir is a corepresentation (i.e., a "glued-up" irrep), timereversal must be set to true; conversely, if lgir is a pseudoreal or complex irrep that would otherwise pair up under time-reversal, timereversal must be set to false.

degree: if provided, monomial terms up to degree will be considered; if degree = nothing (default), the search over monomial terms will terminate after finding the lowest-degree allowed (i.e., nonvanishing) monomial term.

Output

Hs: a vector of MonomialHamiltonian, with each element Hᴹ containing the allowable terms of a k⋅p expansion of incrementing monomial degree degree(Hᴹ) consistent with the transformation properties dictated by lgir.

Each such allowable term Hᴹ $= H^M(\mathbf{k})$ is an expansion in a set of basis elements $H_a^M(\mathbf{k})$:

\[H^M(\mathbf{k}) = \sum_a q_a H_a^M(\mathbf{k})\]

where $q_a$ are free coefficients and $H_a^M(\mathbf{k})$ are the basis elements of the degree $M =$ degree(Hᴹ) monomial terms of the k⋅p Hamiltonian.

The coordinates of the k⋅p model are displayed as x, y, & z for brevity; in practice, they refer to the relative coordinates k₁, k₂, & k₃, respectively, for the k-point k = ∑ᵢkᵢGᵢ, with Gᵢ denoting a reciprocal basis vector. This reciprocal basis is implicitly specified through the choice of coordinates used in specifying the little group operations associated with lgir.

To evaluate Hᴹ for a specific set of expansion vectors at a particular k-point (measured relative to the k-point in lgir, and, as described above, assumed referred to the basis system assumed in lgir) and for a particular set of expansion coefficients qs $= [q_1, q_2, \ldots, q_N]$, Hᴹ can be called as a functor using the syntax Hᴹ(k, qs).

KdotP.Monomial — Type

Monomial{D}(ps::NTuple{D, Int})
Monomial(p1, p2, ..., pD)

Define the D-dimensional monomial

\[x^M = x_1^p_1 \cdots x_i^{p_i} \cdots x_D^{p_D} = \prod_i^D x_i^{p_i}\]

with non-negative integer powers $p_i =$ ps[i]. The degree of the monomial M is $\prod_i p_i$ (see degree(::Monomial)).

A Monomial xᴹ behaves as a functor and can be evaluated at a given position xs = [x₁, …, xᵢ, …, x_D] via xᴹ(xs).

Examples

julia> xᴹ¹ = Monomial(1,2,3)
x₁x₂²x₃³

julia> v¹ = xᴹ¹([1,2,3]) # = 1 * 2^2 * 3^3
108

julia> xᴹ² = Monomial(3,1,0,2)
x₁³x₂x₄²

julia> v² = xᴹ²([1,2,3,5]) # = 1^3 * 2^1 * 3^0 * 5^2
50

KdotP.MonomialBasis — Type

MonomialBasis{D} <: AbstractVector{Monomial}

- `M :: Int` (degree of monomial basis elements)
- `psv :: Vector{Monomomial{D}` (basis elements)

A wrapper around a vector of Monomial{D}, the elements of which constitute a monomial basis for polynomials of degree M in D dimensions.

KdotP.MonomialHamiltonian — Type

MonomialHamiltonian{D}

Fields

hs :: Vector{Hermitian{ComplexF64, Matrix{ComplexF64}}}

bᴹ :: MonomialBasis{D}

cs :: Vector{Vector{Vector{Float64}}}

Description

A MonomialHamiltonian Hᴹ parameterizes the form of a k⋅p Hamiltonian of monomial k-dependence of degree $M =$ degree(Hᴹ):

\[H^M(\mathbf{k}) = \sum_a q_a^M H_a^M(\mathbf{k})\]

where $q_a^M$ are free (expansion) coefficients and $H_a^M(\mathbf{k})$ are the basis elements of the allowed k⋅p Hamiltonian terms of monomial degree $M$.

A MonomialHamiltonian contains the basis elements $\{H_a(\mathbf{k})\}$, allowing only monomial terms of degree $M$ in momentum. In particular, the data structure for the fixed, k-dependent matrices $H_a(\mathbrm{k})$ are of the form:

$H_a(\mathbf{k}) = \sum_n$ H.hs[n] $\sum_j$ H.cs[a][j][n] $k_j^M$

with H denoting an instance of the type MonomialHamiltonian and $k_j^M$ denoting a basis element of the monomials of degree M (see also KdotP.Monomial and KdotP.MonomialBasis).

KdotP.degree — Function

degree(H::Monomial{D}) --> Int

For a monomial

\[x^M = x_1^p_1 \cdots x_i^{p_i} \cdots x_D^{p_D} = \prod_i^D x_i^{p_i}\]

return its degree $M = \prod_i^D p_i$.

degree(H::MonomialBasis) --> Int

Return the degree $M$ of a monomial basis $\{x^M\}$.

See also degree(::Monomial).

degree(H::MonomialHamiltonian) --> Int

Return the degree of the H, i.e., the order or overall exponent of the k-polynomial in H.

See also degree(::MonomialBasis).

degree(Hs::HamiltonianExpansion) --> Int

Return the maximum degree of the MonomialHamiltonians contained in Hs.

See also degree(::MonomialHamiltonian).