KdotP.jl

API

KdotP.kdotp — Function

kdotp(lgir::LGIrrep,
      αβγ=nothing;
      timereversal::Bool = true,
      degree::Union{Integer, Nothing} = nothing)  -->  Vector{MonomialHamiltonian}

Return a basis for the allowed k⋅p Hamiltonians of a small irrep lgir up to linear degree in momentum.

Input arguments

lgir: a small irrep of a little group, provided as an instance of the type LGIrrep. Tables of LGIrreps are accessible from the lgirreps function of Crystalline.jl package.

αβγ: if the little group associated with lgir has free parameters, i.e., if the k-point is parametrized by free parameters (α, β, γ), these parameters may be set via αβγ = [α, β, γ]. If αβγ = nothing (default), lgir is implicitly evaluated at αβγ = [0, 0, 0].

Keyword arguments

timereversal (keyword argument): if true (default), time-reversal invariance is imposed on the k⋅p expansion. If lgir is a corepresentation (i.e., a "glued-up" irrep), timereversal must be set to true; conversely, if lgir is a pseudoreal or complex irrep that would otherwise pair up under time-reversal, timereversal must be set to false.

degree: if provided, monomial terms up to degree will be considered; if degree = nothing (default), the search over monomial terms will terminate after finding the lowest-degree allowed (i.e., nonvanishing) monomial term.

Output

Hs: a vector of MonomialHamiltonian, with each element Hᴹ containing the allowable terms of a k⋅p expansion of incrementing monomial degree degree(Hᴹ) consistent with the transformation properties dictated by lgir.

Each such allowable term Hᴹ $= H^M(\mathbf{k})$ is an expansion in a set of basis elements $H_a^M(\mathbf{k})$:

\[H^M(\mathbf{k}) = \sum_a q_a H_a^M(\mathbf{k})\]

where $q_a$ are free coefficients and $H_a^M(\mathbf{k})$are the basis elements of the degree $M =$ degree(Hᴹ) monomial terms of the k⋅p Hamiltonian. To evaluate Hᴹ for a specific set of expansion vectors at a particular k-point (measured relative to the k-point in lgir, and assumed referred to the basis system assumed in lgir; i.e., generally relative to a reciprocal lattice basis) and for a particular set of expansion coefficients qs $= [q_1, q_2, \ldots, q_N]$, Hᴹ can be called as a functor using the syntax Hᴹ(k, qs).

Missing docstring.

Missing docstring for KPHamiltonian. Check Documenter's build log for details.